renamed a couple of files because they made no sense
